Commit 08.10.21 v1.0

+++ app/routes/atOrder.js
@@ -65,11 +65,11 @@ router.get(
 
             // Проверка прав
             const journals = await jfunction.permissionSet(user);
-            const allowed = journals.find(j => j.id == idJournalName);
-            if(!allowed) return res.status(500)
+            
+            const journal = journals.find(j => j.id == idJournalName); // получаем обект журнала.
+            if(!journal) return res.status(500)
                 .json({errors: ['У тебя нет прав на получение данных этого журнала. Обратись а администатору.'], message: defaultError});
             // Конец проверки прав.
-
             const transactions = await db.executeRequest(`
                 select T.ID,
                     cast(T.DATE_ADDED as date) as DATE_ADDED,
@@ -82,7 +82,7 @@ router.get(
                 left join SALARY_TRANSACTION T on (E.ID_TRANSACTION = T.ID)
                 left join COST_OF_WORK W on (W.ID_JOURNAL = J.ID)
                 left join ORDERS O on (J.ID_ORDER = O.ID)
-                where J.ID_JOURNAL_NAMES = ${idJournalName}
+                where J.ID_JOURNAL_NAMES in (${journal.j.join(', ')})
                 group by T.ID, T.DATE_ADDED, T.NAME
                 order by T.DATE_ADDED desc
             `);
@@ -173,6 +173,10 @@ router.get(
     '/preliminary-calculation/:id',
     async (req, res) => {
         const defaultError = `При получаении данных по предварительному просчету, произошла ошибка.`;
+        const permissions = [
+            'Journals [preliminary-calculation] get all', 
+            'Journals [preliminary-calculation] edit all works'
+        ]
         const idJournalName =  req.params.id;
         try {
             // Проверка токена, получение пользователя.
@@ -182,14 +186,12 @@ router.get(
             catch (error) {return res.status(500).json({errors: [error.message], message: 'Ошибка авторизации.'})}
             const user = await users.getUserToID(decoded.userId);
             // Конец проверки токена.
-
             // Проверка прав
-            const journals = await jfunction.permissionSet(user);
-            const allowed = journals.find(j => j.id == idJournalName);
-            if(!allowed) return res.status(500)
-                .json({errors: ['У тебя нет прав на получение данных этого журнала. Обратись а администатору.'], message: defaultError});
+            const journals = await jfunction.permissionSet(user); // доступные журналы.
+            const journal = journals.find(j => j.id == idJournalName);
+            if(!journal) return res.status(500)
+                .json({errors: ['У тебя нет прав на получение данных этого журнала. Обратись к администатору.'], message: defaultError});
             // Конец проверки прав.
-
             const query = `select O.ID, O.ITM_ORDERNUM, J.ID as ID_JOIRNAL, P.ID_SECTOR, S.NAME as SECTOR, 
                             W.ID as ID_WORK_OF_COST, P.ID as ID_WORK, P.NAME as WORK_NAME,
                             cast(coalesce(O.ORDER_FASADSQ, 0) as decimal(6,3)) as ORDER_FASADSQ, 
@@ -201,10 +203,12 @@ router.get(
                                 left join WORK_PRICES P on (W.ID_WORK_PRICE = P.ID)
                                 left join SECTORS S on (P.ID_SECTOR = S.ID)
                                 left join SALARY_TRANS_EL E on (E.ID_JOURNAL = J.ID)
-                                where E.ID is null and j.id_journal_names = ${idJournalName}`;
+                                where ${await user.permissionCompare(permissions[0]) ? '' : '(P.OPTIONAL != -1) and'}
+                                E.ID is null and j.id_journal_names in (${journal.j.join(', ')})`;
+            const isCanEditAllWorks = await user.permissionCompare(permissions[1]);
             const salary = await db.executeRequest(query);
             const sectors = [];
-            if (salary.length == 0) return res.status(500).json({errors: ['Array empty'], message: 'Список пуст'});
+            if (salary.length == 0) return res.status(500).json({errors: ['В текущем периоде, по этому журналу нет принятых заказов.'], message: 'Список пуст'});
             const sectorsName = [...new Set(salary.map(s => s.SECTOR))];
             for (const sectorName of sectorsName) {
                 const ordersId  = [...new Set(salary.filter(o => o.SECTOR == sectorName).map(o => o.ID))];
@@ -214,6 +218,8 @@ router.get(
                     const idJournal = salary.find(s => s.ID == id).ID_JOIRNAL;
                     const itmOrderNum = salary.find(s => s.ID == id).ITM_ORDERNUM;
                     const works = salary.filter(o => o.ID == id).map(w => {
+                        let checkOptional = w.OPTIONAL;
+                        if(isCanEditAllWorks) checkOptional = -1;
                         return {
                             workOfCostId:   w.ID_WORK_OF_COST,
                             workId:         w.ID_WORK,
@@ -221,10 +227,12 @@ router.get(
                             square:         w.ORDER_FASADSQ,
                             price:          w.PRICE,
                             money:          w.MONEY,
-                            optional:       w.OPTIONAL,
-                            isDeleted:      0
+                            optional:       checkOptional,
+                            isDeleted:      0,
+                            isEdited:       0
                         }
                     });
+
                     const order = {id, itmOrderNum, idJournal, isDeleted:0, works};
                     sector.orders.push(order);
                 }
@@ -265,6 +273,7 @@ router.patch(
         catch (error) {return res.status(500).json({errors: [error.message], message: 'Некорректный токен'})}
         const user = await users.getUserToID(decoded.userId);
         try {
+            
             const transaction = await db.executeRequest(`
                                 insert into SALARY_TRANSACTION (
                                 ID_USER, DATE_ADDED, SETTLEMENT_DATE, NAME, 
@@ -276,26 +285,29 @@ router.patch(
                                     '${format(new Date(), 'DD.MM.YYYY')}',
                                     'Расчет заработной платы', 0, 0, 0, 0
                                 ) returning ID;`
-            )
+            );
             // Обработка работ.
             for (const sector of sectors) {
                 for (const order of sector.orders) {
                     if (order.isDeleted) continue;
                     for (const work of order.works) {
                         if (work.isDeleted) {
-                            await db.executeRequest(`delete from cost_of_work w where w.id = ${work.workOfCostId}`);
+                            if (work.isEdited) await db.executeRequest(`delete from cost_of_work w where w.id = ${work.workOfCostId}`);
                             continue;
                         }else if (work.workOfCostId) {
-                            await db.executeRequest(`update COST_OF_WORK W set W.PRICE = ${work.price} where W.ID = ${work.workOfCostId}`)
+                            await db.executeRequest(`update COST_OF_WORK W set W.PRICE = ${work.price} where W.ID = ${work.workOfCostId}`);
+                            console.log(`Обновили работы (${work.work}) ID ${work.workOfCostId}`);
                         }else {
                             await db.executeRequest(`insert into COST_OF_WORK (ID_JOURNAL, ID_WORK_PRICE, PRICE) values (${order.idJournal}, ${work.workId}, ${work.price})`);
                         }
                     }
-                    await db.executeRequest(`
-                        insert into SALARY_TRANS_EL (ID_JOURNAL, ID_TRANSACTION)
-                        values (${order.idJournal}, ${transaction.ID})
-                    `);
                 }
+                await db.executeRequest(`
+                    insert into SALARY_TRANS_EL (ID_JOURNAL, ID_TRANSACTION)
+                    select j.id, ${transaction.ID}
+                    from journals j
+                    where j.id in (${sector.orders.filter(o => !o.isDeleted).map(o => o.idJournal).join(', ')})
+                `)
                  // Доп списания / начисления
                 if (sector.otherTransactoins.data.length > 0) {
                     for (const otherTransaction of sector.otherTransactoins.data) {
@@ -336,15 +348,16 @@ router.post(
                         .json({errors: errors.array(), message: defaultError});
 
             const {idTransfer: transferBarcode, idAccepted: acceptedBarcode, orders: registerOrders, ...other} = req.body;
-            console.log(other);
 
             if (!Array.isArray(registerOrders) || registerOrders.length <= 0) return res.status(500)
                         .json({errors: ['Нет заказов для передачи.'], message: defaultError});
             if (transferBarcode == acceptedBarcode) return res.status(500)
                         .json({errors: ['Нельзя передавать заказ самому себе.'], message: defaultError});
+
             let query = atOrderQuery.get('get_barcodes', {
                 $where: `UPPER(B.BARCODE) = '${transferBarcode.toUpperCase()}' 
                             or UPPER(B.BARCODE) = '${acceptedBarcode.toUpperCase()}'`});
+
             const barcodes = await db.executeRequest(query);
             const transfer = barcodes.find(item => item.BARCODE.toUpperCase() === transferBarcode.toUpperCase());
             const accepted = barcodes.find(item => item.BARCODE.toUpperCase() === acceptedBarcode.toUpperCase());
@@ -358,7 +371,7 @@ router.post(
 
             const namesTransferOldSector  = await getNameOldSectorArrToIdNewSector(transfer.ID_SECTOR);
             const namesAcceptedOldSector  = await getNameOldSectorArrToIdNewSector(accepted.ID_SECTOR); 
-            
+
             query = atOrderQuery.get('get_dep', {
                     $where: `
                     D.ID_SECTOR_TRANSFER = ${transfer ? transfer.ID_SECTOR : null} and
@@ -370,16 +383,13 @@ router.post(
                                 '"отправитель" не определен и'} не может передавать заказы ${accepted ? 'участку ' + accepted.SECTOR : 
                                         'не определенному участку.'}`], 
                         message: defaultError});
-
             query = orderQuery.get('get_orders', {$where: `O.ID IN (${registerOrders.map(o => o.idOrder).join(', ')})`});    
             const orders = await db.executeRequest(query);
-
             query = `select J.ID, J.ID_ORDER, N.NAME
                         from JOURNALS J
                         left join JOURNAL_NAMES N on (J.ID_JOURNAL_NAMES = N.ID)
                         where J.ID_JOURNAL_NAMES in (${journal.map(j => j.ID_JOURNAL_NAME).join(', ')}) and
                         J.ID_ORDER in (${registerOrders.map(o => o.idOrder).join(', ')})`;
-
             const adoptedOrders = await db.executeRequest(query);
             for (const o of registerOrders) {
                 const order = orders.find(ord => ord.ID === o.idOrder);
@@ -401,23 +411,20 @@ router.post(
                         o.description = `Заказ уже принят в "${isAdopted.NAME}"`;
                         continue;
                     }
+
                     const statusAllow = journal.find(j => j.STATUS_NUM === order.ORDER_STATUS)
-                    if (!statusAllow) {
+
+                    const allowStatuses = journal.filter(j => !!j.STATUS_NUM); // Есть ли требуемые статусы
+
+
+                    if (!statusAllow && (allowStatuses && allowStatuses.length)) {
                         o.completed = false;
                         o.description = `Не верный статус "${order.STATUS_DESCRIPTION}", ожидается: "${journal.map(j => j.STATUS_DESCRIPTION).join(', ')}"`;
                         continue;
                     }
                     o.completed = true;
                     o.description = `успешно!`;
-                    // Открываем транзакцию
-                    db.beginTransaction(async (err, trans) => {
-                        if (err) {
-                            o.completed = false;
-                            o.description = `Ошибка сохранения в базу данных`;
-                            console.log('Transaction error', err);
-                            return;
-                        }
-                        const query = `
+                    const query = `
                                 execute block
                                 returns (ID integer)
                                 as
@@ -425,6 +432,7 @@ router.post(
                                     insert into JOURNALS (ID_ORDER, ID_JOURNAL_NAMES, NOTE, TS, TRANSFER_DATE)
                                     values (${order.ID}, ${journal[0].ID_JOURNAL_NAME}, ${o.comment ? '\'' + o.comment + '\'' : null}, 
                                         '${format(new Date(), 'DD.MM.YYYY HH:mm:ss')}', '${format(other.date, 'DD.MM.YYYY HH:mm:ss')}') returning ID into :ID;
+
                                     insert into JOURNAL_TRANS (ID_EMPLOYEE, ID_SECTOR, ID_JOURNAL, MODIFER)
                                     values (${transfer.ID_EMPLOYEE}, ${transfer.ID_SECTOR}, :ID, -1);
                                     insert into JOURNAL_TRANS (ID_EMPLOYEE, ID_SECTOR, ID_JOURNAL, MODIFER)
@@ -434,51 +442,9 @@ router.post(
                                     where p.id_sector = ${transfer.ID_SECTOR};
                                     suspend;
                                 end`;
-                        //Выполнение запроса на добавление записи в журнал
-                        trans.query(query, async (err, result) => {
-                            if (err) {
-                                trans.rollback();
-                                o.completed = false;
-                                o.description = `Ошибка сохранения в базу данных`;
-                                console.log('Transaction error', err);
-                                return;
-                            }
-                            const newStatus = await getStatusNumOldToIdStatusOld(journal[0].ID_STATUS_AFTER);
-                            if (newStatus) {
-                                //Есди назначен новый статус, выполняем запрос на обновление статуса.
-                                trans.query(
-                                    `update ORDERS O set O.ORDER_STATUS = ${newStatus} where O.ID = ${order.ID}`, 
-                                    async (err, result) => {
-                                        if (err) {
-                                            //если ошибка, откатываем транзакцию.
-                                            trans.rollback();
-                                            o.completed = false;
-                                            o.description = `Ошибка сохранения в базу данных`;
-                                            console.log('Transaction error', err);
-                                            return;
-                                        }
-                                        // Подтверждаем транзакцию
-                                        trans.commit((err) => {
-                                            if (err) {
-                                                o.completed = false;
-                                                o.description = `Ошибка транзакции.`;
-                                                trans.rollback();
-                                            }
-                                        });
-                                    }
-                                )
-                            }else{
-                                //Подтверждаем транзакцию.
-                                trans.commit((err) => {
-                                    if (err) {
-                                        o.completed = false;
-                                        o.description = `Ошибка транзакции.`;
-                                        trans.rollback();
-                                    }
-                                });
-                            }
-                        })    
-                    });
+                    await db.executeRequest(query);
+                    const newStatus = await getStatusNumOldToIdStatusOld(journal[0].ID_STATUS_AFTER);
+                    if (newStatus) await db.executeRequest(`update ORDERS O set O.ORDER_STATUS = ${newStatus} where O.ID = ${order.ID}`);
                 }else{
                     o.completed = false;
                     o.description = `Номер заказа [${o.idOrder}] не найден в базе данных.`;
@@ -486,7 +452,7 @@ router.post(
             }
             const countOrders = registerOrders.filter(o => o.completed).length;
             let resultMessage = `${countOrders ? '☑️ Принято ' + countOrders + ' из ' + registerOrders.length : '❌ Не один из заказов не принят.'}`;
-            if (countOrders == registerOrders.length)  resultMessage = '✅ Все заказы приняты.'
+            if (countOrders == registerOrders.length)  resultMessage = `✅ Все заказы приняты. (${registerOrders.length})`
             return res.status(201).json({
                 message: resultMessage,
                 orders: registerOrders
